chore: Move player data error handler to player data port file
